Finished lab 10

+++ lab10/server/film-dao.mjs
@@ -11,10 +11,10 @@ export function FilmLibrary(){
     /*
     a. Retrieve all the stored films and return a Promise that resolves to an array of Film objects. 
    */
-    this.getFilms = () => {
+    this.getFilms = (userId) => {
         return new Promise((resolve, reject) =>{
-            const sql = "SELECT * FROM films";
-            db.all(sql, (err, rows) =>{
+            const sql = "SELECT * FROM films WHERE userId = ?";
+            db.all(sql, [userId], (err, rows) =>{
                 if(err){
                     reject(err);
                 }else if(rows !== undefined){
@@ -29,10 +29,10 @@ export function FilmLibrary(){
    /*
     b. Retrieve all favorite films and return a Promise that resolves to an array of Film objects. 
    */
-    this.getFavoriteFilms = () => {
+    this.getFavoriteFilms = (userId) => {
         return new Promise((resolve, reject) => {
-            const sql = "SELECT * FROM films WHERE isFavorite = True";
-            db.all(sql, (err, rows) => {
+            const sql = "SELECT * FROM films WHERE isFavorite = True AND userId = ?";
+            db.all(sql, [userId], (err, rows) => {
                 if(err){
                     reject(err);
                 }else if(rows !== undefined){
@@ -47,10 +47,10 @@ export function FilmLibrary(){
    /*
     c. Retrieve all films watched today and return a Promise that resolves to an array of Film objects. 
    */
-  this.getFilmsWatchedToday = () => {
+  this.getFilmsWatchedToday = (userId) => {
     return new Promise((resolve, reject) => {
-        const sql = "SELECT * FROM films WHERE watchDate = ?";
-        db.all(sql, [dayjs().format("YYYY-MM-DD")], (err, rows) => {
+        const sql = "SELECT * FROM films WHERE watchDate = ? AND userId = ?";
+        db.all(sql, [dayjs().format("YYYY-MM-DD"), userId], (err, rows) => {
             if(err){
                 reject(err);
             }else if(rows !== undefined){
@@ -65,10 +65,10 @@ export function FilmLibrary(){
   /*
     c. Retrieve all films watched last month and return a Promise that resolves to an array of Film objects. 
    */
-    this.getFilmsWatchedLastMonth = () => {
+    this.getFilmsWatchedLastMonth = (userId) => {
       return new Promise((resolve, reject) => {
-          const sql = "SELECT * FROM films WHERE watchDate >= ?";
-          db.all(sql, [dayjs().startOf('month').format("YYYY-MM-DD")], (err, rows) => {
+          const sql = "SELECT * FROM films WHERE watchDate >= ? AND userId = ?";
+          db.all(sql, [dayjs().startOf('month').format("YYYY-MM-DD"), userId], (err, rows) => {
               if(err){
                   reject(err);
               }else if(rows !== undefined){
@@ -84,10 +84,10 @@ export function FilmLibrary(){
     d. Retrieve films whose watch date is earlier than a given date (received as a parameter). 
        Return a Promise that resolves to an array of Film objects. 
    */
-  this.getFilmsPreviousWatched = (date) => {
+  this.getFilmsPreviousWatched = (date, userId) => {
     return new Promise((resolve, reject) => {
-        const sql = "SELECT * FROM films WHERE watchDate < ?";
-        db.all(sql, [date.format("YYYY-MM-DD")], (err, rows) => {
+        const sql = "SELECT * FROM films WHERE watchDate < ? AND userId = ?";
+        db.all(sql, [date.format("YYYY-MM-DD"), userId], (err, rows) => {
             if(err){
                 reject(err);
             }else if(rows !== undefined){
@@ -103,10 +103,10 @@ export function FilmLibrary(){
     d. Retrieve films never watched
        Return a Promise that resolves to an array of Film objects. 
    */
-       this.getUnseenFilms = (date) => {
+       this.getUnseenFilms = (userId) => {
         return new Promise((resolve, reject) => {
-            const sql = "SELECT * FROM films WHERE watchDate IS NULL";
-            db.all(sql, (err, rows) => {
+            const sql = "SELECT * FROM films WHERE watchDate IS NULL AND userId = ?";
+            db.all(sql, [userId], (err, rows) => {
                 if(err){
                     reject(err);
                 }else if(rows !== undefined){
@@ -122,10 +122,10 @@ export function FilmLibrary(){
     e. Retrieve films whose rating is greater than or equal to a given number (received as a parameter).
        Return a Promise that resolves to an array of Film objects. 
    */
-  this.getFilmsWithHigherRateThan = (rate) => {
+  this.getFilmsWithHigherRateThan = (rate, userId) => {
     return new Promise((resolve, reject) => {
-        const sql = "SELECT * FROM films WHERE rating > ?";
-        db.all(sql, [rate], (err, rows) => {
+        const sql = "SELECT * FROM films WHERE rating > ? AND userId = ?";
+        db.all(sql, [rate, userId], (err, rows) => {
             if(err){
                 reject(err);
             }else if(rows !== undefined){
@@ -140,11 +140,11 @@ export function FilmLibrary(){
     f. Retrieve films whose title contains a given string (received as a parameter). 
        Return a Promise that resolves to an array of Film objects. 
    */
-  this.getFilmsTitleContains = (keyWord) => {
+  this.getFilmsTitleContains = (keyWord, userId) => {
     keyWord = `%${keyWord.toLowerCase()}%`;
     return new Promise((resolve, reject) => {
-        const sql = "SELECT * FROM films WHERE title LIKE ?";
-        db.all(sql, [keyWord], (err, rows) => {
+        const sql = "SELECT * FROM films WHERE title LIKE ? AND userId = ?";
+        db.all(sql, [keyWord, userId], (err, rows) => {
             if(err){
                 reject(err);
             }else if(rows !== undefined){
@@ -180,8 +180,8 @@ export function FilmLibrary(){
   */
   this.addFilm = (film) => {
     return new Promise((resolve, reject) => {
-      let sql = 'SELECT * FROM films WHERE title = ?';
-      db.get(sql, [film.title], (err, row) => {
+      let sql = 'SELECT * FROM films WHERE title = ? AND userId = ?';
+      db.get(sql, [film.title, film.userId], (err, row) => {
         if (err) {
           reject(err);
         } else if (row === undefined) {
@@ -205,12 +205,12 @@ export function FilmLibrary(){
   */
   this.updateFilm = (film) => {
     return new Promise((resolve, reject) => {
-      let sql = 'SELECT * FROM films WHERE id = ?';
-      db.get(sql, [film.id], (err, row) => {
+      let sql = 'SELECT * FROM films WHERE id = ? AND userId = ?';
+      db.get(sql, [film.id, film.userId], (err, row) => {
         if (err) {
           reject(err);
         } else if (row !== undefined) {
-          sql = 'UPDATE films SET title = ?, isFavorite = ?, watchDate = DATE(?), rating = ? WHERE id = ?';
+          sql = 'UPDATE films SET title = ?, isFavorite = ?, watchDate = DATE(?), rating = ? WHERE id = ? AND userId = ?';
           
           if(film.title === undefined)
             film.title = row.title;
@@ -221,7 +221,7 @@ export function FilmLibrary(){
           if(film.rating === undefined)
             film.rating = row.rating;
           
-          db.run(sql, [film.title, film.isFavorite, film.watchDate, film.rating, film.id], (err) => {
+          db.run(sql, [film.title, film.isFavorite, film.watchDate, film.rating, film.id, film.userId], (err) => {
             if (err){
               reject(err);
             }else{
@@ -241,13 +241,13 @@ export function FilmLibrary(){
   */
 this.deleteFilm = (film) => {
     return new Promise((resolve, reject) => {
-      let sql = 'SELECT * FROM films WHERE title = ?';
-      db.get(sql, [film.title], (err, row) => {
+      let sql = 'SELECT * FROM films WHERE title = ? AND userId = ?';
+      db.get(sql, [film.title, film.userId], (err, row) => {
         if (err) {
           reject(err);
         } else if (row !== undefined) {
-          sql = 'DELETE FROM films WHERE title = ?';
-          db.run(sql, [film.title], (err) => {
+          sql = 'DELETE FROM films WHERE title = ? AND userId = ?';
+          db.run(sql, [film.title, film.userId], (err) => {
             if (err){
               reject(err);
             }else{
@@ -265,15 +265,15 @@ this.deleteFilm = (film) => {
   b. Delete a movie from the database (using its ID as a reference). 
      After completion, print a confirmation/failure message. 
   */
-this.deleteFilmById = (id) => {
+this.deleteFilmById = (id, userId) => {
   return new Promise((resolve, reject) => {
-    let sql = 'SELECT * FROM films WHERE id = ?';
-    db.get(sql, [id], (err, row) => {
+    let sql = 'SELECT * FROM films WHERE id = ? AND userId = ?';
+    db.get(sql, [id, userId], (err, row) => {
       if (err) {
         reject(err);
       } else if (row !== undefined) {
-        sql = 'DELETE FROM films WHERE id = ?';
-        db.run(sql, [id], (err) => {
+        sql = 'DELETE FROM films WHERE id = ? AND userId = ?';
+        db.run(sql, [id, userId], (err) => {
           if (err){
             reject(err);
           }else{
@@ -291,10 +291,10 @@ this.deleteFilmById = (id) => {
   c. Delete the watch date of all films stored in the database. 
      After completion, print a confirmation/failure message. 
   */
- this.deleteWathcedDates = () => {
+ this.deleteWathcedDates = (userId) => {
     return new Promise((resolve, reject) => {
-      let sql = 'UPDATE films SET watchDate = null';
-      db.run(sql, (err) => {
+      let sql = 'UPDATE films SET watchDate = null WHERE userId = ?';
+      db.run(sql, [userId], (err) => {
         if (err){
             reject(err);
         }else{
